SDG-5505: wifi: mt76: fix queue assignment for action frames
When running in AP mode and sending an action frame to a client that's
in powersave mode, the frame can get stuck in a tx queue along with
other buffered frames. This can fill up hardware queues with frames
that are only released after the WTBL slot is reused for another client.

Fix this by moving action frames to the ALTX queue.

Also add debugfs files to dump token usage and content.