SDG-5505: wifi: mt76: fix queue assignment for EAPOL frames
When running in AP mode and sending an EAPOL frame to a client that's
in powersave mode, the frame can get stuck in a tx queue along with
other buffered frames. This can fill up hardware queues with frames
that are only released after the WTBL slot is reused for another client.

Fix this by moving EAPOL frames to the ALTX queue.

Signed-off-by: Chad Monroe <chad@monroe.io>